+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 
 - There is only one configuration file: `bbmri.negotiator.xml` which includes all previously existing configuration settings.
   Copy the `bbmri.negotiator.xml.example` file to `bbmri.negotiator.xml` and configure it to your needs.
+- New configuration values:
+    - proxy settings
+    - molgenis settings (URL, resource names)